finish the node and media CRUD in central

+++ central/src/main/java/com/brendhacasaro/digital_media/node/NodeRepository.java
@@ -3,4 +3,4 @@ package com.brendhacasaro.digital_media.node;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface NodeRepository extends JpaRepository<Node, Long> {}
+public interface NodeRepository extends JpaRepository<Node, Integer> {}
